backend/requirements.txt
fastapi
uvicorn
requests
biopython
llama-index
llama-index-vector-stores-chroma
chromadb
llama-index-llms-gemini
llama-index-embeddings-gemini
google-generativeai
google-cloud-aiplatform
python-dotenv
pandas
